refactor: Replace ExecutionContext impls with ne... (#25)
Github-Change-Id: 956579
GitOrigin-RevId: 01b77b800d62f1eef7fab69844fcbbe1f3cbd97c
