+++ tenant/runtime/src/integrationTest/kotlin/viaduct/tenant/runtime/context/MutationFieldExecutionContextImplTest.kt
@@ -1,44 +1,59 @@
-@file:Suppress("ForbiddenImport")
+@file:OptIn(ExperimentalCoroutinesApi::class)
 
 package viaduct.tenant.runtime.context
 
-import io.mockk.every
 import io.mockk.mockk
-import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.runBlockingTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import viaduct.api.internal.select.SelectionsLoader
+import viaduct.api.mocks.MockGlobalIDCodec
+import viaduct.api.mocks.MockInternalContext
 import viaduct.api.select.SelectionSet
-import viaduct.api.types.Arguments
-import viaduct.api.types.CompositeOutput
-import viaduct.api.types.Mutation
-import viaduct.api.types.Object
-import viaduct.api.types.Query
-import viaduct.tenant.runtime.select.Mutation as SelectMutation
+import viaduct.tenant.runtime.select.Mutation
+import viaduct.tenant.runtime.select.SelectTestFeatureAppTest
 
-class MutationFieldExecutionContextImplTest {
-    private val mutationObject = mockk<SelectMutation>()
+class MutationFieldExecutionContextImplTest : ContextTestBase() {
+    private val mutationObject = mockk<Mutation>()
 
-    private fun mk(
-        inner: FieldExecutionContextImpl<*, *, *, *> = mockk<FieldExecutionContextImpl<Object, Query, Arguments, CompositeOutput>>(),
-        mutationLoader: SelectionsLoader<Mutation> = SelectionsLoader.const(mutationObject)
-    ) = MutationFieldExecutionContextImpl(
-        inner,
-        mutationLoader
-    )
+    private fun mk(): MutationFieldExecutionContextImpl {
+        val wrapper = createMockingWrapper(
+            schema = SelectTestFeatureAppTest.schema,
+            mutationMock = mutationObject
+        )
+
+        return MutationFieldExecutionContextImpl(
+            MockInternalContext(SelectTestFeatureAppTest.schema, MockGlobalIDCodec()),
+            wrapper,
+            noSelections,
+            null, // requestContext
+            Args,
+            Obj,
+            Q,
+        )
+    }
 
     @Test
-    fun mutation(): Unit =
-        runBlocking {
+    fun mutation() =
+        runBlockingTest {
             val ctx = mk()
-            assertEquals(mutationObject, ctx.mutation(SelectionSet.empty(SelectMutation.Reflection)))
+            assertEquals(mutationObject, ctx.mutation(SelectionSet.empty(Mutation.Reflection)))
         }
 
     @Test
     fun delegation() {
-        val inner = mockk<FieldExecutionContextImpl<Object, Query, Arguments, CompositeOutput>>()
-        every { inner.arguments }.returns(Arguments.NoArguments)
-        val ctx = mk(inner)
-        assertEquals(Arguments.NoArguments, ctx.arguments)
+        val ctx = mk()
+        // Test that basic properties are accessible (delegation works)
+        assertEquals(Args, ctx.arguments)
+        assertEquals(Obj, ctx.objectValue)
+        assertEquals(Q, ctx.queryValue)
+        assertEquals(SelectionSet.NoSelections, ctx.selections())
+    }
+
+    @Test
+    fun selectionsFor() {
+        val ctx = mk()
+        // Test that selectionsFor works for mutations (no exception thrown)
+        ctx.selectionsFor(Mutation.Reflection, "__typename")
     }
 }
