+++ tenant/runtime/src/main/kotlin/viaduct/tenant/runtime/context/MutationFieldExecutionContextImpl.kt
@@ -1,9 +1,7 @@
 package viaduct.tenant.runtime.context
 
-import viaduct.api.context.FieldExecutionContext
 import viaduct.api.context.MutationFieldExecutionContext
 import viaduct.api.internal.InternalContext
-import viaduct.api.internal.select.SelectionsLoader
 import viaduct.api.select.SelectionSet
 import viaduct.api.types.Arguments
 import viaduct.api.types.CompositeOutput
@@ -11,9 +9,20 @@ import viaduct.api.types.Mutation
 import viaduct.api.types.Object
 import viaduct.api.types.Query
 
-class MutationFieldExecutionContextImpl<T : Object, Q : Query, A : Arguments, O : CompositeOutput>(
-    private val inner: FieldExecutionContextImpl<T, Q, A, O>,
-    private val mutationLoader: SelectionsLoader<Mutation>,
-) : FieldExecutionContext<T, Q, A, O> by inner, MutationFieldExecutionContext<T, Q, A, O>, InternalContext by inner {
-    override suspend fun <T : Mutation> mutation(selections: SelectionSet<T>): T = mutationLoader.load(this, selections)
+/**
+ * Note the primary constructor is for testing purposes, to allow you
+ * to more easily mock out the engine execution context.  Production
+ * use cases should use the secondary constructor.
+ */
+class MutationFieldExecutionContextImpl(
+    baseData: InternalContext,
+    engineExecutionContextWrapper: EngineExecutionContextWrapper,
+    selections: SelectionSet<CompositeOutput>,
+    requestContext: Any?,
+    arguments: Arguments,
+    objectValue: Object,
+    queryValue: Query,
+) : MutationFieldExecutionContext<Object, Query, Arguments, CompositeOutput>,
+    SealedFieldExecutionContextImpl(baseData, engineExecutionContextWrapper, selections, requestContext, arguments, objectValue, queryValue) {
+    override suspend fun <T : Mutation> mutation(selections: SelectionSet<T>) = engineExecutionContextWrapper.mutation(this, "mutation", selections)
 }
