--- tenant/runtime/src/main/kotlin/viaduct/tenant/runtime/context2/VariablesProviderContextImpl.kt
@@ -1,11 +1,0 @@
-package viaduct.tenant.runtime.context2
-
-import viaduct.api.context.VariablesProviderContext
-import viaduct.api.internal.InternalContext
-import viaduct.api.types.Arguments
-
-class VariablesProviderContextImpl<A : Arguments>(
-    baseData: InternalContext,
-    override val requestContext: Any?,
-    override val arguments: A
-) : VariablesProviderContext<A>, ExecutionContextImpl(baseData)
