+++ tenant/runtime/src/main/kotlin/viaduct/tenant/runtime/execution/FieldBatchResolverExecutorImpl.kt
@@ -15,7 +15,7 @@ import viaduct.engine.api.FieldResolverExecutor
 import viaduct.engine.api.FieldResolverExecutor.Selector
 import viaduct.engine.api.RequiredSelectionSet
 import viaduct.engine.api.ResolverMetadata
-import viaduct.tenant.runtime.context2.factory.FieldExecutionContextFactory
+import viaduct.tenant.runtime.context.factory.FieldExecutionContextFactory
 
 /**
  * Executes a tenant-written field resolver's batchResolve function.
